Set home page CMS to work with new data structure
