+++ src/views/cms/components/TheSideBar.vue
@@ -1,0 +1,11 @@
+<template>
+  <div class="sidebar">
+    <router-link to="/home">Home</router-link>
+    <router-link to="/learn">Learn</router-link>
+  </div>
+</template>
+<script>
+export default {
+  name: "TheSideBar"
+};
+</script>
