+++ src/views/cms/views/home/components/FourthSection.vue
@@ -1,42 +1,42 @@
 <template>
-  <b-container class="team-section">
-    <h2 class="gordita-medium">Footer Section Content</h2>
+  <b-container class="hero-section">
+    <h2 class="gordita-medium">Fourth Section Content</h2>
     <form>
       <b-row class="mt-5">
-        <b-col md="4">
+        <b-col>
+          <div class="form-section">
+            <label>Leading to Main Branded text</label>
+            <textarea v-model="sectionData.leadingToMainText"></textarea>
+          </div>
+        </b-col>
+        <b-col>
           <div class="form-section">
             <label>Main Branded text</label>
             <textarea v-model="sectionData.mainText"></textarea>
           </div>
         </b-col>
-        <b-col md="4">
+        <b-col>
           <div class="form-section">
-            <label>Survey Text</label>
-            <textarea v-model="sectionData.surveyText"></textarea>
-          </div>
-        </b-col>
-        <b-col md="4">
-          <div class="form-section">
-            <label>Address</label>
-            <textarea v-model="sectionData.address"></textarea>
+            <label>Main Sub Text</label>
+            <textarea v-model="sectionData.mainSubText"></textarea>
           </div>
         </b-col>
       </b-row>
       <b-row class="mt-5">
-        <b-col md="4">
+        <b-col md="6">
           <div class="form-section">
-            <label>Copyright Text</label>
-            <textarea v-model="sectionData.copyrightText"></textarea>
+            <label>Quote Text</label>
+            <textarea v-model="sectionData.quoteText"></textarea>
           </div>
         </b-col>
       </b-row>
-      <button @click.prevent="updateSection(false)" class="mt-5 ml-1">
+      <button @click.prevent="updateSection(false)" class="mt-5">
         <fai icon="upload" />
-        Update Footer Section's Content
+        Update Fourth Section's Content
       </button>
       <button @click.prevent="updateSection(true)" class="mt-5 ml-1">
         <fai icon="magic" />
-        Preview Changes to Footer Section
+        Preview Changes to Fourth Section
       </button>
     </form>
     <EmailAlert v-if="showAlert" :type="type" :content="content" />
@@ -44,10 +44,10 @@
 </template>
 
 <script>
-import mixin from "../mixins/network_update_content";
+import mixin from "@/views/cms/mixins/network_update_content";
 import EmailAlert from "@/views/home/components/EmailAlert";
 export default {
-  name: "FooterSection",
+  name: "FourthSection",
   components: { EmailAlert },
   mixins: [mixin],
   data: () => ({
@@ -55,26 +55,26 @@ export default {
     type: "success",
     content: "",
     sectionData: {
+      leadingToMainText: "",
       mainText: "",
-      surveyText: "",
-      address: "",
-      copyrightText: ""
+      mainSubText: "",
+      quoteText: ""
     }
   }),
-  mounted() {
-    Object.keys(this.sectionData).forEach(key => {
-      this.sectionData[key] = this.websiteData.footer[key];
-    });
-  },
   methods: {
     updateSection: async function(update = false) {
       let currentData = await this.getWebsiteData(this.websiteData, update);
-      currentData.footer = this.sectionData;
+      currentData.homePage.fourthSection = this.sectionData;
       if (update)
         this.previewContent(currentData, "Your changes will be previewed");
       else
-        this.updateContent(currentData, "Footer Section Updated Successfully");
+        this.updateContent(currentData, "Fourth Section Updated Successfully");
     }
+  },
+  mounted() {
+    Object.keys(this.sectionData).forEach(key => {
+      this.sectionData[key] = this.websiteData.fourthSection[key];
+    });
   },
   props: {
     websiteData: Object
